requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
pandas>=2.0.0
Pillow>=10.0.0

# Optional: PP-Structure (PaddleOCR) local document/table structure engine.
# Install these when you want to use `engine=ppstructure` or set `OCR_ENGINE=ppstructure`.
# NOTE: On Windows, prefer installing `paddlepaddle` from official instructions (CPU/GPU variants).
# paddleocr>=2.7.0.3
# paddlepaddle>=2.6.0
# lxml>=5.0.0
